refactor: exporting the used routes directly from the individual pages
