+++ src/pages/Index/IndexPage.tsx
@@ -1,12 +1,18 @@
-import AppShell from "../components/app-shell";
-import Protector from "../routes/Protector";
-import StatisticsWidget from "../components/Dashboard/statistics-widget";
-import BannerNotice from "../components/Dashboard/banner-notice";
-import { useGetUserProfile } from "../hooks/network/useGetUserProfile";
-import { useGetDashboardStats } from "../hooks/network/useGetDashboardStats";
-import { useGetDashboardNoticeList } from "../hooks/network/useGetDashboardNoticeList";
+import { rootRoute } from "../../routes/Router";
+import AppShell from "../../components/app-shell";
+import Protector from "../../routes/Protector";
+import StatisticsWidget from "../../components/Dashboard/statistics-widget";
+import BannerNotice from "../../components/Dashboard/banner-notice";
+import { useGetUserProfile } from "../../hooks/network/useGetUserProfile";
+import { useGetDashboardStats } from "../../hooks/network/useGetDashboardStats";
+import { useGetDashboardNoticeList } from "../../hooks/network/useGetDashboardNoticeList";
 
-function Index() {
+export const indexRoute = rootRoute.createRoute({
+  path: "/",
+  component: IndexPage,
+});
+
+function IndexPage() {
   const userQuery = useGetUserProfile();
   const statistics = useGetDashboardStats({
     locationId: 0,
@@ -45,4 +51,4 @@ function Index() {
   );
 }
 
-export default Index;
+export default IndexPage;
