+++ src/routes/Router.tsx
@@ -5,13 +5,12 @@ import {
   parseSearchWith,
   stringifySearchWith,
 } from "@tanstack/react-router";
-import { z } from "zod";
 
 import { OidcAuthProvider } from "./OidcAuthProvider";
-import IndexPage from "../pages/Index";
-import AgreementsPage from "../pages/Agreements";
+import { indexRoute } from "../pages/Index/IndexPage";
+import { agreementsSearchRoute } from "../pages/AgreementsSearch/AgreementsSearchPage";
 
-const rootRoute = createRouteConfig({
+export const rootRoute = createRouteConfig({
   component: () => {
     return (
       <OidcAuthProvider>
@@ -19,40 +18,6 @@ const rootRoute = createRouteConfig({
       </OidcAuthProvider>
     );
   },
-});
-
-const indexRoute = rootRoute.createRoute({
-  path: "/",
-  component: IndexPage,
-});
-
-const agreementFiltersModel = z
-  .object({
-    AgreementStatusName: z.string().optional(),
-    Statuses: z.number().optional(),
-    IsSearchOverdues: z.boolean().optional(),
-    EndDate: z
-      .preprocess((arg) => {
-        if (typeof arg == "string" || arg instanceof Date) return new Date(arg);
-      }, z.date())
-      .optional(),
-  })
-  .optional();
-
-export const agreementsSearchRoute = rootRoute.createRoute({
-  path: "/agreements",
-  component: AgreementsPage,
-  validateSearch: z.object({
-    page: z.number().min(1).default(1),
-    size: z.number().min(1).default(10),
-    filters: agreementFiltersModel,
-  }).parse,
-  preSearchFilters: [
-    (search) => ({
-      page: search.page || 1,
-      size: search.size || 10,
-    }),
-  ],
 });
 
 const routeConfig = rootRoute.addChildren([indexRoute, agreementsSearchRoute]);
